requirements.txt
numpy==1.24.3
scipy==1.11.1
opencv-python==4.8.0.74
PyWavelets==1.4.1
Flask==2.3.2
requests==2.31.0
pillow==10.0.0
mss==9.0.1
pynput==1.7.6
cryptography==41.0.3

